+++ bin/functions.sh
@@ -21,6 +21,7 @@ wpdl-codesniff() {
 	wpdl-codesniff-php-syntax
 	wpdl-codesniff-php-autoloaders
 	wpdl-codesniff-phpcs
+	wpdl-codesniff-strings
 	wpdl-codesniff-l10n
 	wpdl-codesniff-bash
 	wpdl-codesniff-jshint
@@ -76,18 +77,47 @@ wpdl-codesniff-php-autoloaders() {
 	fi
 }
 
-# Check php files with PHPCodeSniffer.
-wpdl-codesniff-phpcs() {
-	local path=$(wpdl-get-codesniff-path PHP PHPCS)
-	local files=$(find "${!path}")
+# Check php files with PHPCodeSniffer tools.
+wpdl-codesniff-phpcs-base() {
+
+	local command=${1-phpcs}
+
+	if [ -z $2 ]; then
+		local path=$(wpdl-get-codesniff-path PHP PHPCS)
+		local files=$(find "${!path}")
+	else
+		local files=("$2")
+	fi
 
 	if [ ! -e $PHPCS_DIR ]; then
-		local phpcs=phpcs
+		local phpcs="$command"
 	else
-		local phpcs="$PHPCS_DIR"/scripts/phpcs
+		local phpcs="$PHPCS_DIR"/scripts/"$command"
 	fi
 
 	"$phpcs" -ns --standard="$WPCS_STANDARD" ${files[@]}
+}
+
+# Check php files with PHPCS.
+wpdl-codesniff-phpcs() {
+	wpdl-codesniff-phpcs-base phpcs "${@}"
+}
+
+# Check php files with PHP Can Be Fixed.
+wpdl-codesniff-phpcbf() {
+	wpdl-codesniff-phpcs-base phpcbf "${@}"
+}
+
+# Check files for disallowed strings.
+wpdl-codesniff-strings() {
+
+	local path=$(wpdl-get-codesniff-path STRINGS)
+	local files=$(find "${!path}" -type f)
+
+	grep -n -v "${CODESNIFF_IGNORED_STRINGS[@]}" ${files[@]} | grep -e 'target="_blank"' -e http[^s_.-]
+
+	# grep exits with 1 if nothing was found.
+	[[ $? == '1' ]]
 }
 
 # Check JS files with jshint.
@@ -163,7 +193,9 @@ wpdl-test-phpunit() {
 		GROUP_OPTION=(--group="$TEST_GROUP")
 		CLOVER_FILE+="-$TEST_GROUP"
 
-		if [[ $TRAVIS_PHP_VERSION == '5.2' ]]; then
+		if [[ $TEST_GROUP == uninstall && -e phpunit.uninstall.xml.dist ]]; then
+			GROUP_OPTION=(--configuration=phpunit.uninstall.xml.dist)
+		elif [[ $TRAVIS_PHP_VERSION == '5.2' ]]; then
 			sed -i '' -e "s/<group>$TEST_GROUP<\/group>//" ./phpunit.xml.dist
 		fi
 	fi