+++ phpunit/module-bootstrap.php
@@ -33,12 +33,25 @@ define( 'RUNNING_WORDPOINTS_MODULE_TESTS', true );
 define( 'WORDPOINTS_MODULE_TESTS_DIR', dirname( dirname( dirname( __FILE__ ) ) ) . '/tests/phpunit' );
 
 /**
+ * The class autoloader for the PHPUnit helper classes.
+ *
+ * @since 2.4.0
+ */
+require_once( dirname( __FILE__ ) . '/classes/class/autoloader.php' );
+
+WordPoints_Dev_Lib_PHPUnit_Class_Autoloader::register_dir(
+	dirname( __FILE__ ) . '/classes/'
+	, 'WordPoints_Dev_Lib_PHPUnit_'
+);
+
+/**
  * The bootstrap's utility functions.
  *
  * @since 1.1.0
  */
 require_once( dirname( __FILE__ ) . '/functions.php' );
 
+// This is mainly left here for back-compat with pre 2.5.0 behavior.
 $has_uninstall_tester = is_dir( WORDPOINTS_MODULE_TESTS_DIR .  '/../../vendor/wordpoints/module-uninstall-tester/' );
 
 if ( $has_uninstall_tester ) {
@@ -61,12 +74,44 @@ if ( $has_uninstall_tester ) {
 /**
  * The WordPress tests functions.
  *
- * We are loading this so that we can add our tests filter to load the module and
- * WordPoints, using tests_add_filter().
+ * We are loading this for legacy reasons. We used to have to include it to load the
+ * module and WordPoints using tests_add_filter().
  *
  * @since 1.0.0
  */
 require_once( getenv( 'WP_TESTS_DIR' ) . '/includes/functions.php');
+
+if ( file_exists( WORDPOINTS_MODULE_TESTS_DIR .  '/../../vendor/autoload_52.php' ) ) {
+
+	/**
+	 * The PHP 5.2 compatible Composer autoloader for the module's dependencies.
+	 *
+	 * @since 2.5.0
+	 */
+	require_once( WORDPOINTS_MODULE_TESTS_DIR .  '/../../vendor/autoload_52.php' );
+
+} elseif ( file_exists( WORDPOINTS_MODULE_TESTS_DIR .  '/../../vendor/autoload.php' ) ) {
+
+	/**
+	 * The Composer generated autoloader for the module's dependencies.
+	 *
+	 * @since 2.5.0
+	 */
+	require_once( WORDPOINTS_MODULE_TESTS_DIR .  '/../../vendor/autoload.php' );
+}
+
+if (
+	class_exists( 'WPPPB_Loader' )
+	&& file_exists( getenv( 'WORDPOINTS_TESTS_DIR' ) . '/classes/bootstrap/loader.php' )
+) {
+
+	/**
+	 * WordPoints's PHPUnit loader class.
+	 *
+	 * @since 2.5.0
+	 */
+	require_once( getenv( 'WORDPOINTS_TESTS_DIR' ) . '/classes/bootstrap/loader.php' );
+}
 
 if ( file_exists( WORDPOINTS_MODULE_TESTS_DIR . '/includes/functions.php' ) ) {
 
@@ -78,7 +123,18 @@ if ( file_exists( WORDPOINTS_MODULE_TESTS_DIR . '/includes/functions.php' ) ) {
 	require_once( WORDPOINTS_MODULE_TESTS_DIR . '/includes/functions.php' );
 }
 
-if ( ! $has_uninstall_tester || ! running_wordpoints_module_uninstall_tests() ) {
+if (
+	class_exists( 'WordPoints_PHPUnit_Bootstrap_Loader' )
+	&& ! defined( 'WORDPOINTS_MODULE_TESTS_LOADER' )
+) {
+
+	$loader = WordPoints_PHPUnit_Bootstrap_Loader::instance();
+	$loader->add_module(
+		wordpoints_dev_lib_the_module_basename()
+		, getenv( 'WORDPOINTS_MODULE_NETWORK_ACTIVE' )
+	);
+
+} elseif ( ! $has_uninstall_tester || ! running_wordpoints_module_uninstall_tests() ) {
 
 	// Hook to load WordPoints.
 	tests_add_filter( 'muplugins_loaded', 'wordpointstests_manually_load_plugin' );
@@ -127,12 +183,19 @@ if ( $has_uninstall_tester ) {
 	require_once( WORDPOINTS_MODULE_TESTS_DIR . '/../../vendor/wordpoints/module-uninstall-tester/bootstrap.php' );
 }
 
-/**
- * The class autoloader for the PHPUnit helper classes.
- *
- * @since 2.4.0
- */
-require_once( dirname( __FILE__ ) . '/classes/class/autoloader.php' );
+if (
+	( ! $has_uninstall_tester || ! running_wordpoints_module_uninstall_tests() )
+	&& ( ! isset( $loader ) || ! $loader->running_uninstall_tests() )
+	&& file_exists( WORDPOINTS_MODULE_TESTS_DIR . '/../../src/admin/admin.php' )
+) {
+
+	/**
+	 * The module's admin-side code.
+	 *
+	 * @since 1.0.0
+	 */
+	require_once( WORDPOINTS_MODULE_TESTS_DIR . '/../../src/admin/admin.php' );
+}
 
 if ( file_exists( WORDPOINTS_MODULE_TESTS_DIR . '/includes/bootstrap.php' ) ) {