+++ wpcept/includes/Modules/WordPointsLoader.php
@@ -9,6 +9,7 @@
 
 namespace WordPoints\Tests\Codeception\Modules;
 
+use WordPoints\Tests\Codeception\BlackHoleCache;
 use Codeception\Module;
 use Codeception\Configuration;
 use Codeception\Exception\ModuleException;
@@ -62,10 +63,8 @@ class WordPointsLoader extends Module {
 		// Now get a dump of the pristine database so that we can restore it later.
 		$this->create_db_dump( $this->get_db_dump_file_name() );
 
-		// Flush the cache and suspend caching, since the DB can be modified during
-		// the tests.
+		// Suspend caching, since the DB can be modified during the tests.
 		$this->suspend_caching();
-		$this->flush_cache();
 	}
 
 	/**
@@ -85,6 +84,16 @@ class WordPointsLoader extends Module {
 				__CLASS__
 				, "\nWP_TESTS_DIR is not set."
 			);
+		}
+
+		// We indicate whether to run as multisite via a file.
+		// Properly handling for this must be added to the site's wp-config.php.
+		$multisite_file = getenv( 'WP_TESTS_DIR' ) . '/../../is-multisite';
+
+		if ( getenv( 'WP_MULTISITE' ) ) {
+			touch( $multisite_file );
+		} elseif ( file_exists( $multisite_file ) ) {
+			unlink( $multisite_file );
 		}
 
 		// Catch output from PHPUnit bootstrap.
@@ -116,7 +125,11 @@ class WordPointsLoader extends Module {
 	 */
 	protected function load_wordpoints() {
 
-		$result = activate_plugin( 'wordpoints/wordpoints.php' );
+		$result = activate_plugin(
+			'wordpoints/wordpoints.php'
+			, ''
+			, (bool) getenv( 'WORDPOINTS_NETWORK_ACTIVE' )
+		);
 
 		if ( is_wp_error( $result ) ) {
 			throw new ModuleException(
@@ -144,7 +157,11 @@ class WordPointsLoader extends Module {
 	 */
 	protected function load_wordpoints_module( $module ) {
 
-		$result = wordpoints_activate_module( $module );
+		$result = wordpoints_activate_module(
+			$module
+			, ''
+			, (bool) getenv( 'WORDPOINTS_MODULE_NETWORK_ACTIVE' )
+		);
 
 		if ( is_wp_error( $result ) ) {
 			throw new ModuleException(
@@ -243,11 +260,11 @@ class WordPointsLoader extends Module {
 
 		$result = shell_exec(
 			vsprintf(
-				'mysqldump --host=%s -u %s --password=%s %s 2>&1 1> %s'
+				'MYSQL_PWD=%s mysqldump --host=%s -u %s %s 2>&1 1> %s'
 				, array(
+					escapeshellarg( DB_PASSWORD ),
 					escapeshellarg( DB_HOST ),
 					escapeshellarg( DB_USER ),
-					escapeshellarg( DB_PASSWORD ),
 					escapeshellarg( DB_NAME ),
 					escapeshellarg( $dump_file ),
 				)
@@ -310,12 +327,12 @@ class WordPointsLoader extends Module {
 
 		$result = shell_exec(
 			vsprintf(
-				'cat %s | mysql --host=%s -u %s --password=%s %s 2>&1 1> /dev/null'
+				'cat %s | MYSQL_PWD=%s mysql --host=%s -u %s %s 2>&1 1> /dev/null'
 				, array(
 					escapeshellarg( $dump_file ),
+					escapeshellarg( DB_PASSWORD ),
 					escapeshellarg( DB_HOST ),
 					escapeshellarg( DB_USER ),
-					escapeshellarg( DB_PASSWORD ),
 					escapeshellarg( DB_NAME ),
 				)
 			)
@@ -335,15 +352,22 @@ class WordPointsLoader extends Module {
 	 * @since 2.4.0
 	 */
 	protected function suspend_caching() {
-		wp_suspend_cache_addition( true );
+
+		global $wp_object_cache;
+
+		$wp_object_cache = new BlackHoleCache();
 	}
 
 	/**
 	 * Flush the WordPress object cache.
 	 *
 	 * @since 2.4.0
+	 * @deprecated 2.5.0 Now implied by self::suspend_caching().
 	 */
 	protected function flush_cache() {
+
+		_deprecated_function( __FUNCTION__, '2.5.0 of the dev-lib' );
+
 		wp_cache_flush();
 	}
 }